+++ src/js/items/help.js
@@ -1,0 +1,8 @@
+// Config
+const { PREFIX } = require("../../config.json");
+
+module.exports = {
+    title: `Commands list (${PREFIX}command)`,
+    maxFieldsPage: 5
+}
+