+++ src/js/main-functions/create-help.js
@@ -1,4 +1,7 @@
+// Items
+const help = require("../items/help");
 // Nano functions
+const createFieldsHelpMessageEmbed = require("../nano-functions/create-fields-help-message-embed");
 const createMessageEmbed = require("../nano-functions/create-message-embed");
 const addReactionsLastMessage = require ("../nano-functions/add-reactions-last-message");
 const reactInternalError = require("../nano-functions/react-internal-error");
@@ -8,17 +11,10 @@ const _ = require("lodash");
 
 module.exports = (bot, message, commands) => {
     try {
-        let fields = _.cloneDeep(commands);
+        let fields = createFieldsHelpMessageEmbed(commands, 0, help.maxFieldsPage);
+        let footer = '1/' + Math.ceil(commands.length/fields.length);
 
-        fields.forEach((field) => {
-            field.name = field.commands.toString().replace(',', ', ');
-            delete field.commands;
-            field.value = field.description + (field.arguments ? '\n ..................................... \n' + field.arguments : '');
-            delete field.description;
-            delete field.arguments
-        });
-
-        message.channel.send(createMessageEmbed(`Commands list (${PREFIX}command)`, fields));
+        message.channel.send(createMessageEmbed(help.title, fields, footer));
 
         _.delay(() => addReactionsLastMessage(message, ['⬅', '➡']), 200);
         _.delay(() => message.delete(), 200);