+++ src/js/main-functions/dont-allow-auto-read-reaction.js
@@ -1,0 +1,5 @@
+module.exports = (messageReaction) => {
+    if (messageReaction.users.cache.last().bot) return false;
+
+    return true;
+}